Added tests and documentation for redux-promise middleware
