+++ docs/README.md
@@ -14,6 +14,7 @@
   * [Middleware](/docs/advanced/middleware/README.md)
     * [redux-thunk](/docs/advanced/middleware/redux-thunk.md)
     * [redux-saga](/docs/advanced/middleware/redux-saga.md)
+    * [redux-promise](/docs/advanced/middleware/redux-promise.md)
     * [Custom Middleware](/docs/advanced/middleware/CustomMiddleware.md)
 * [API Reference](/docs/api/README.md)
   * [subspace](/docs/api/subspace.md)
