+++ docs/advanced/README.md
@@ -6,4 +6,5 @@
 * [Middleware](/docs/advanced/middleware/README.md)
   * [redux-thunk](/docs/advanced/middleware/redux-thunk.md)
   * [redux-saga](/docs/advanced/middleware/redux-saga.md)
+  * [redux-promise](/docs/advanced/middleware/redux-promise.md)
   * [Custom Middleware](/docs/advanced/middleware/CustomMiddleware.md)
